Move bool caster tests to where they belong
